requirements.txt
fastapi
uvicorn
python-dotenv
pillow
numpy
matplotlib
opencv-python
pytest
httpx
torch
torchvision
segmentation-models-pytorch
albumentations
scikit-learn
python-multipart
